backend/requirements.txt
fastapi
uvicorn
pyserial
requests
supabase
python-dotenv
python-socketio[asgi]
PyJWT
jinja2
email-validator
